LITS

LITS

API documentation

LITS

Index

Exported

Internal

Index

LITS.AVRSimple

LITS.AndersonFouadMachine

LITS.AvgCnvFixedDC

LITS.CombinedVIwithVZ

LITS.DirectControl

LITS.DynGenerator

LITS.LCFilter

LITS.MarconatoMachine

LITS.OneDOneQMachine

LITS.PLL

LITS.ReactivePowerDroop

LITS.SimpleAFMachine

LITS.SimpleMarconatoMachine

LITS.StaticSource

LITS.VirtualInertia

Exported

LITS.AVRSimple — Type.

Parameters of a simple proportional AVR in the derivative of EMF i.e. an integrator controller on EMF

LITS.AndersonFouadMachine — Type.

Parameters of 6-states synchronous machine: Anderson-Fouad model

Conmutable structor

AndersonFouadMachine(R, Xd, Xq, Xd_p, Xq_p, Xd_pp, Xq_pp, Td0_p, Tq0_p, Td0_pp, Tq0_pp, MVABase)

#Arguments

R::Float64 : Resistance after EMF in machine per unit

Xd::Float64 : Reactance after EMF in d-axis per unit

Xq::Float64 : Reactance after EMF in q-axis per unit

Xd_p::Float64 : Transient reactance after EMF in d-axis per unit

Xq_p::Float64 : Transient reactance after EMF in q-axis per unit

Xd_pp::Float64 : Subtransient reactance after EMF in d-axis per unit

Xq_pp::Float64 : Subtransient reactance after EMF in q-axis per unit

Td0_p::Float64 : Time constant of transient d-axis voltage

Tq0_p::Float64 : Time constant of transient q-axis voltage

Td0_pp::Float64 : Time constant of subtransient d-axis voltage

Tq0_pp::Float64 : Time constant of subtransient q-axis voltage

MVABase::Float64 : Nominal Capacity in MVA

LITS.AvgCnvFixedDC — Type.

v_rated::Float64 : rated voltage

s_rated::Float64 : rated VA

LITS.CombinedVIwithVZ — Type.

kpc::Float64 : current controller proportional gain

kic::Float64 : current controller integral gain

kpv::Float64 : voltage controller proportional gain

kiv::Float64 : voltage controller integral gain

ω_ad::Float64 : active damping filter cutoff frequency (rad/sec)

kad::Float64 : active damping gain

lv::Float64 : virtual inductance

rv::Float64 : virtual resistance

LITS.DynGenerator — Type.

LITS.MarconatoMachine — Type.

Parameters of 6-states synchronous machine: Marconato model

Conmutable structor

MarconatoMachine(R, Xd, Xq, Xd_p, Xq_p, Xd_pp, Xq_pp, Td0_p, Tq0_p, Td0_pp, Tq0_pp, T_AA, MVABase)

#Arguments

R::Float64 : Resistance after EMF in machine per unit

Xd::Float64 : Reactance after EMF in d-axis per unit

Xq::Float64 : Reactance after EMF in q-axis per unit

Xd_p::Float64 : Transient reactance after EMF in d-axis per unit

Xq_p::Float64 : Transient reactance after EMF in q-axis per unit

Xd_pp::Float64 : Subtransient reactance after EMF in d-axis per unit

Xq_pp::Float64 : Subtransient reactance after EMF in q-axis per unit

Td0_p::Float64 : Time constant of transient d-axis voltage

Tq0_p::Float64 : Time constant of transient q-axis voltage

Td0_pp::Float64 : Time constant of subtransient d-axis voltage

Tq0_pp::Float64 : Time constant of subtransient q-axis voltage

T_AA::Float64 : Time constant of d-axis additional leakage

MVABase::Float64 : Nominal Capacity in MVA

LITS.OneDOneQMachine — Type.

Parameters of 2-states synchronous machine: One d- and One q-Axis Model

Conmutable structor

OneDOneQMachine(R, Xd, Xq, Xd_p, Xq_pp, Td0_p, Tq0_pp, MVABase)

#Arguments

R::Float64

Xd::Float64

Xq::Float64

Xd_p::Float64 : Reactance after EMF in machine per unit

Xq_pp::Float64

Td0_p::Float64 : Time constant of transient d-axis voltage

Tq0_pp::Float64 : Time constant of subtransient q-axis voltage

MVABase::Float64 : Nominal Capacity in MVA

LITS.PLL — Type.

ω_lp::Float64 : PLL low-pass filter frequency (rad/sec)

kp_pll::Float64 : PLL proportional gain

ki_pll::Float64 : PLL integral gain

LITS.ReactivePowerDroop — Type.

kq::Float64 : reactive power droop gain

ωf::Float64 : reactive power filter cutoff frequency (rad/sec)

LITS.SimpleAFMachine — Type.

Parameters of 4-states simplified Anderson-Fouad (SimpleAFMachine) model. The derivative of stator fluxes (ψd and ψq) is neglected and ωψd = ψd and ωψq = ψq is assumed (i.e. ω=1.0). This is standard when transmission network dynamics is neglected. If transmission dynamics is considered use the full order Anderson Fouad model.

Conmutable structor

SimpleAFMachine(R, Xd, Xq, Xd_p, Xq_p, Xd_pp, Xq_pp, Td0_p, Tq0_p, Td0_pp, Tq0_pp, MVABase)

#Arguments

R::Float64 : Resistance after EMF in machine per unit

Xd::Float64 : Reactance after EMF in d-axis per unit

Xq::Float64 : Reactance after EMF in q-axis per unit

Xd_p::Float64 : Transient reactance after EMF in d-axis per unit

Xq_p::Float64 : Transient reactance after EMF in q-axis per unit

Xd_pp::Float64 : Subtransient reactance after EMF in d-axis per unit

Xq_pp::Float64 : Subtransient reactance after EMF in q-axis per unit

Td0_p::Float64 : Time constant of transient d-axis voltage

Tq0_p::Float64 : Time constant of transient q-axis voltage

Td0_pp::Float64 : Time constant of subtransient d-axis voltage

Tq0_pp::Float64 : Time constant of subtransient q-axis voltage

MVABase::Float64 : Nominal Capacity in MVA

LITS.SimpleMarconatoMachine — Type.

Parameters of 4-states synchronous machine: Simplified Marconato model The derivative of stator fluxes (ψd and ψq) is neglected and ωψd = ψd and ωψq = ψq is assumed (i.e. ω=1.0). This is standard when transmission network dynamics is neglected.

Conmutable structor

SimpleMarconatoMachine(R, Xd, Xq, Xd_p, Xq_p, Xd_pp, Xq_pp, Td0_p, Tq0_p, Td0_pp, Tq0_pp, T_AA, MVABase)

#Arguments

R::Float64 : Resistance after EMF in machine per unit

Xd::Float64 : Reactance after EMF in d-axis per unit

Xq::Float64 : Reactance after EMF in q-axis per unit

Xd_p::Float64 : Transient reactance after EMF in d-axis per unit

Xq_p::Float64 : Transient reactance after EMF in q-axis per unit

Xd_pp::Float64 : Subtransient reactance after EMF in d-axis per unit

Xq_pp::Float64 : Subtransient reactance after EMF in q-axis per unit

Td0_p::Float64 : Time constant of transient d-axis voltage

Tq0_p::Float64 : Time constant of transient q-axis voltage

Td0_pp::Float64 : Time constant of subtransient d-axis voltage

Tq0_pp::Float64 : Time constant of subtransient q-axis voltage

T_AA::Float64 : Time constant of d-axis additional leakage

MVABase::Float64 : Nominal Capacity in MVA

LITS.StaticSource — Type.

This struct acts as an infinity bus.

LITS.VirtualInertia — Type.

Ta::Float64 : VSM interia constant

kd::Float64 : VSM damping constant

kw::Float64 : frequency droop gain

ωb::Float64 : rated angular frequency

Internal

LITS.DirectControl — Type.

kpc::Float64 : current controller proportional gain

kic::Float64 : current controller integral gain

kpv::Float64 : voltage controller proportional gain

kiv::Float64 : voltage controller integral gain

ω_ad::Float64 : active damping filter cutoff frequency (rad/sec)

kad::Float64 : active damping gain

lv::Float64 : virtual inductance

rv::Float64 : virtual resistance

LITS.LCFilter — Type.

lf::Float64 : filter inductance

rf::Float64 : filter resistance

cf::Float64 : filter capacitance