LITS

API documentation

LITS

Index

Exported

Internal

Index

LITS.configure_logging

LITS.get_initial_conditions

LITS.get_voltagemag_series

LITS.initialize_mach_shaft!

LITS.initialize_mach_shaft!

LITS.initialize_mach_shaft!

LITS.initialize_mach_shaft!

LITS.initialize_mach_shaft!

LITS.initialize_mach_shaft!

LITS.kirchoff_laws!

LITS.print_device_states

LITS.print_init_states

Exported

LITS.get_initial_conditions — Method

Returns a Dictionary with the resulting initial conditions of the simulation

LITS.get_voltagemag_series — Method

Function to obtain the voltage magnitude series out of the DAE Solution. It receives the solution, the dynamical system and the bus number.

LITS.print_device_states — Method

Function to print initial states. It receives the vector of initial states and the dynamical system.

LITS.print_init_states — Method

Function to print initial states. It receives the vector of initial states and the dynamical system.

Internal

LITS.configure_logging — Method

configure_logging(;
    console_level = Logging.Error,
    file_level = Logging.Info,
    filename = "power-simulations.log",
)

Creates console and file loggers.

Note: Log messages may not be written to the file until flush() or close() is called on the returned logger.

Arguments

console_level = Logging.Error: level for console messages

file_level = Logging.Info: level for file messages

filename::String = power-simulations.log: log file

Example

logger = configure_logging(console_level = Logging.Info)
@info "log message"
close(logger)

LITS.initialize_mach_shaft! — Method

Initialitation of model of 6-state (Anderson-Fouad) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

LITS.initialize_mach_shaft! — Method

Initialitation of model of 0-state synchronous (classic model) machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

LITS.initialize_mach_shaft! — Method

Initialitation of model of 6-state (Marconato) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

LITS.initialize_mach_shaft! — Method

Initialitation of model of 2-state (One d- and One q-) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

LITS.initialize_mach_shaft! — Method

Initialitation of model of 4-state (Simple Anderson-Fouad) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

LITS.initialize_mach_shaft! — Method

Initialitation of model of 4-state (Simple Marconato) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

LITS.kirchoff_laws! — Method

Kirchoff law for buses.
I_gen_r[j]: Real current injection from all generators connected at bus j.
            It is zero if no generator is connected at bus j.
I_gen_i[j]: Real current injection from all generators connected at bus j.
            It is zero if no generator is connected at bus j.